@@ -10,9 +10,19 @@
         "preview": "vite preview"
     },
     "dependencies": {
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-router-dom": "^5.3.4"
+        "@headlessui/react": "^2.2.4",
+        "@heroicons/react": "^2.2.0",
+        "@tailwindcss/forms": "^0.5.10",
+        "autoprefixer": "^10.4.16",
+        "axios": "^1.10.0",
+        "jwt-decode": "^4.0.0",
+        "postcss": "^8.4.31",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "react-hook-form": "^7.60.0",
+        "react-router-dom": "^6.30.1",
+        "react-toastify": "^11.0.5",
+        "tailwindcss": "^3.3.5"
     },
     "devDependencies": {
         "@types/react": "^18.2.56",
