@@ -8,18 +8,7 @@ import {
   updateField,
   addHealthRecord,
   getNearbyFields
-} from '../controllers/field.controller.js';utes/fieldRoutes.js
-import express from 'express';
-import { protect } from '../middleware/authMiddleware.js';
-import { upload } from '../config/cloudinary.js';
-import {
-  createField,
-  getFields,
-  getField,
-  updateField,
-  addHealthRecord,
-  getNearbyFields
-} from '../controllers/fieldController.js';
+} from '../controllers/field.controller.js';
 
 const router = express.Router();
 
