@@ -1,6 +1,6 @@
 import express from 'express';
-import { getCurrentUser } from '../controllers/user.js';
-import {  authMiddleware } from '../middleware/authMiddleware.js';
+import { getCurrentUser } from '../controllers/user.controller.js';
+import {  authMiddleware } from '../middleware/auth.middleware.js';
 
 const router = express.Router();
 
